completed schedulemaker gui, added folder with test schedules
The employee schedulemaker gui should be fully functional now.  added a folder with test schedules. Note: i believe the test13.csv file is the only one with a full schedule filled out, the rest are incomplete or don't have the right format and should generate an error when trying to open them.