Calculator, Day, Person, Form1, Schedule Creator
Calculator: It now Parses one Particular schedule (test 13, the working
one). I need to test in on a not uniformly filled schedule still and
extend to parse all schedules in a folder.

Day: Added a ToString Method for Debug

Person: Modified ToString. Added ConsoleDebug for Debug

Form1: This is just for testing. Not permanent.

Schedule Creator: Temporarly linked "Calculate Schedule" to Form1 for
testing.