Calculator Main Logic
Main Logic is mostly done!

Calculator has an update to Calculate()
Person and Job have new functions to help this CanDo() and
CanDoHelper(). They are written in such a way as to preserve "proper"
object oriented design.

I still need to make sure I am taking into account the maximum hours
each worker should work. That will be done in the AddJob function, which
does a lot of the conflict testing. I'll probably have to rewrite the
person constructor too. Will add to the UI if needed.