Lots of Stuff :P
Debugging is good so far. I added a People Schedule Random Generator for
Testing. If we line up the right number of people and jobs, we should
have the whole thing working correctly. Lots of spot checking needs to
be done on results.

The Results Still need to be written out to a file. The results are
currently written out to the Debug Console, if you want to look.

I changed Form1 to now ask for the number of hours each worker should
work. Form 1 probably needs a better name.
It also needs a "You're finished!" pop up box with info on where to find
the file of results.