Almost Done!
I fixed some parsing and generating code for schedules both in the
schedule maker and the calculator function.

The results are now writen to "results.txt" in the same folder as the
sln file. Check it out!

There is also some dialog boxes to help users understand when
calculations are complete.